External Console: Move inputhooks.py to utils
