+++ spyder/utils/inputhooks.py
